// frontend/src/api/models/ModelListItem.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ModelListItem = {
    tool_id: string;
    name: string;
    version: string;
    description: string;
    modality: string;
    task_type: string;
    input_shape: Array<number>;
    class_names: Array<string>;
    loaded: boolean;
};